refactor: MVC 패턴에 따라 모델, 뷰, 컨트롤러 각각의 패키지로 구조 조직화
