+++ src/main/java/baseball/model/Computer.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.model;
 
 import java.util.List;
