+++ src/main/java/baseball/model/Player.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.model;
 
 import java.util.Arrays;
 import java.util.List;
