+++ src/main/java/baseball/view/Output.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.view;
 
 public class Output {
